src/auth/pyproject.toml
[project]
name = "supabase_auth"
version = "2.12.3" # {x-release-please-version}
description = "Python Client Library for Supabase Auth"
authors = [
  {name = "Joel Lee", email = "joel@joellee.org" }
]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

requires-python = ">=3.9"
dependencies = [
  "httpx[http2] >=0.26,<0.29",
  "pydantic >=1.10,<3",
  "pyjwt >=2.10.1",
]

[project.urls]
homepage = "https://github.com/supabase/auth-py"
repository = "https://github.com/supabase/auth-py"
documentation = "https://github.com/supabase/auth-py"

[project.scripts]
gh-download = "scripts.gh-download:main"

[dependency-groups]
tests = [
  "Faker >= 37.4.0",
  "respx >=0.20.2,<0.23.0",
  "pytest >= 8.4.1",
  "pytest-mock >= 3.14.0",
  "pytest-cov >= 6.2.1",
  "pytest-depends >= 1.0.1",
  "pytest-asyncio >= 1.0.0",
]
lints = [
  "ruff >=0.12.1",
  "unasync >= 0.6.0",
]
dev = [{ include-group = "lints" }, {include-group = "tests" }]

[tool.uv]
default-groups = [ "dev" ]

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  # "B",
  # flake8-simplify
  # "SIM",
  # isort
  "I",
]
ignore = ["F401", "F403", "F841", "E712", "E501", "E402", "E722", "E731", "UP006", "UP035"]
# isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true


[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
